mcOnly finished debugging and also updated sim starstar/starplanet bool names accross the board.

+++ code/pythonSide/ProcessManager.py
@@ -371,7 +371,7 @@ def multiProcessStarter(paramSettingsDict):
     ############################################################
     # finish the Gelman-Rubin statistic calculations if requested
     ############################################################
-    if paramSettingsDict['CalcGelmanRubin']and paramSettingsDict['useMultiProcessing']:
+    if (paramSettingsDict['CalcGelmanRubin']and paramSettingsDict['useMultiProcessing'])and(paramSettingsDict["mcONLY"]==False):
         tools.gen.gelmanRubinStage2(dataFiles)    
     
     ####################################################################################################
@@ -397,7 +397,7 @@ def multiProcessStarter(paramSettingsDict):
         origFiles.append(os.path.basename(summaryPlotFile)+"-ChiSquaredDist.png")
         if paramSettingsDict['DIonly']==False:
             origFiles.append(os.path.basename(summaryPlotFile)+"-RVoffsets.png")
-        if paramSettingsDict['removeBurnIn']:
+        if paramSettingsDict['removeBurnIn']and(paramSettingsDict["mcONLY"]==False):
             origFiles.append(os.path.basename(summaryPlotFile)+"-burnInRemoved.png")
             origFiles.append(os.path.basename(summaryPlotFile)+"-burnInRemoved-ChiSquaredDist.png")
             if paramSettingsDict['DIonly']==False:
@@ -429,7 +429,7 @@ def multiProcessStarter(paramSettingsDict):
             for chainNum in range(1,len(dataFiles)+1):
                 origFiles.append(os.path.basename(simAnnealSummaryRootPlotFilename)+"-chain_"+str(chainNum)+".png")
     # get GR filenames
-    if paramSettingsDict['CalcGelmanRubin']and paramSettingsDict['useMultiProcessing']:
+    if (paramSettingsDict['CalcGelmanRubin']and paramSettingsDict['useMultiProcessing'])and(paramSettingsDict["mcONLY"]==False):
         origFiles.append('GRvalues.txt')
         origFiles.append('Tvalues.txt')
     # get RESULTS.txt filenames
@@ -484,24 +484,25 @@ def multiProcessStarter(paramSettingsDict):
                 print '\nDeleting MCMC chain data files\n'+"-"*40
                 for filename in dataFiles:
                     print 'Deleting file: '+os.path.basename(filename)
-                    os.remove(filename)
-                if paramSettingsDict['removeBurnIn']:
+                    os.remove(filename)     
+                if (paramSettingsDict['removeBurnIn'])and(paramSettingsDict["mcONLY"]==False):
                     print '\nDeleting Burn-In removed MCMC chain data files\n'+"-"*40
                     for filename in strippedNames:
                         print 'Deleting file: '+os.path.basename(filename)
-                        os.remove(filename)
+                        os.remove(filename)           
         else:
             print '\n\nDeleting final output chain data files\n'+"-"*40
             for filename in dataFiles:
                 print 'Deleting file: '+os.path.basename(filename)
                 os.remove(filename)
-            if paramSettingsDict['removeBurnIn']:
-                    print '\nDeleting Burn-In removed MCMC chain data files\n'+"-"*40
-                    for filename in strippedNames:
-                        print 'Deleting file: '+os.path.basename(filename)
-                        os.remove(filename)
+            if (paramSettingsDict['removeBurnIn'])and(paramSettingsDict["mcONLY"]==False):
+                print '\nDeleting Burn-In removed MCMC chain data files\n'+"-"*40
+                for filename in strippedNames:
+                    print 'Deleting file: '+os.path.basename(filename)
+                    os.remove(filename)
+            
     ## delete GR chain files if requested
-    if (paramSettingsDict['delGRchainFiles'] and (len(dataFiles)>1)):
+    if (paramSettingsDict['delGRchainFiles'] and (len(dataFiles)>1))and(paramSettingsDict["mcONLY"]==False):
         print '\n\nDeleting final output GR chain value files\n'+"-"*40
         for chainNum in range(1,len(dataFiles)+1):
             filename = os.path.join(paramSettingsDict['outputData_dir'],"gelmanRubin-chain_"+str(chainNum)+".txt")
@@ -515,7 +516,7 @@ def multiProcessStarter(paramSettingsDict):
         if cleanDataFilename!='':
             print 'Deleting file: '+cleanDataFilename
             os.remove(cleanDataFilename)
-        if paramSettingsDict['removeBurnIn']:
+        if (paramSettingsDict['removeBurnIn'])and(paramSettingsDict["mcONLY"]==False):
             print 'Deleting file: '+dataFinalFilename2
             os.remove(dataFinalFilename2)
